Created new python file for items, separated the items and landing page view, created mini components and added them accordingly, routing is a small issue right now but most things should be working
