+++ application/templates/Components/addButtons/addItemsAndCategory.html
@@ -21,7 +21,7 @@
                     <span aria-hidden="true">&times;</span>
                 </button>
             </div>
-            <form method="POST" action="/">
+            <form method="POST" action="/category">
 
                 <div class="modal-body">
                     {{form.csrf_token}}
