+++ application/templates/Components/listingComponents/listItems.html
@@ -1,0 +1,16 @@
+<section class='itemLists'>
+    <div class="card">
+        <div class="card-header">
+            All Items
+        </div>
+        <ul class="list-group list-group-flush">
+            {%for items in itemList%}
+            <li class="list-group-item">{{items.item_name}}
+                <a href="/items/update/{{items.item_ID}}" class="btn btn-outline-secondary btn-sm">Update</a>
+                <a href="/items/delete/{{items.item_ID}}" class="btn btn-outline-danger btn-sm">Delete</a>
+            </li>
+            {%endfor%}
+        </ul>
+    </div>
+
+</section>
